+++ tests/test_pdf_polynomials.py
@@ -8,9 +8,8 @@ import tensorflow_probability as tfp
 
 import zfit
 # noinspection PyUnresolvedReferences
-from zfit import z
+from zfit.core.testing import setup_function, teardown_function, tester
 
-# obs1_random = zfit.Space(obs="obs1", limits=(-1.05, 1.05))
 obs1_random = zfit.Space(obs="obs1", limits=(-1.5, 1.2))
 obs1 = zfit.Space(obs="obs1", limits=(-1, 1))