+++ zfit/pdf.py
@@ -2,19 +2,21 @@
 
 from .core.basepdf import BasePDF
 from .models.basic import Exponential
-from .models.dist_tfp import Gauss, Uniform, WrapDistribution, TruncatedGauss, Cauchy
+from .models.dist_tfp import Gauss, Uniform, WrapDistribution, TruncatedGauss, Cauchy, Poisson
 from .models.functor import ProductPDF, SumPDF, BaseFunctor
 from .models.kde import GaussianKDE1DimV1
 from .models.physics import CrystalBall, DoubleCB
 from .models.polynomials import Chebyshev, Legendre, Chebyshev2, Hermite, Laguerre, RecursivePolynomial
 from .models.special import ZPDF, SimplePDF, SimpleFunctorPDF
+from .models.convolution import FFTConvPDFV1
 
 __all__ = ['BasePDF', 'BaseFunctor',
            'Exponential',
            'CrystalBall', 'DoubleCB',
-           'Gauss', 'Uniform', 'TruncatedGauss', 'WrapDistribution', 'Cauchy',
+           'Gauss', 'Uniform', 'TruncatedGauss', 'WrapDistribution', 'Cauchy', 'Poisson',
            "Chebyshev", "Legendre", "Chebyshev2", "Hermite", "Laguerre", "RecursivePolynomial",
            'ProductPDF', 'SumPDF',
            'GaussianKDE1DimV1',
+           'FFTConvPDFV1',
            'ZPDF', 'SimplePDF', 'SimpleFunctorPDF'
            ]