+++ zfit/util/deprecation.py
@@ -1,0 +1,3 @@
+#  Copyright (c) 2020 zfit
+
+from .legacy import deprecated  # noqa